window-slot: Always reuse current view
We used to have 2 NautilusView implementations (NautilusFilesView and
NautilusPlacesView) which meant the slot was supposed to switch
between them when going into or out of other-locations:///.

But Other Locations place is gone, and so is NautilusPlacesView. Now
every view is NautilusFilesView. Yet, we still destroy the current
view and create a new one when going into or out of network-view:///.

There is no reason to do so anymore, because NautilusFilesView can
switch view id on the fly using `nautilus_files_view_change()`.

So, take the assumption that NautilusFilesView can handle any view_id
and reuse it every time. Also add a missing property change notification
for NautilusWindowSlot:toolbar-view-menu (it was okay before because
that property doesn't change between grid and list, but it does now
between these and network).