window-slot: Get view_id for location, not view
Now that we reuse the same view for all locations, it doesn't make
sense to get a view for a location. Instead we want the view mode id.

This makes it explicit that we only create a new view when we do not
have any yet, and further simplifies the setup of new locations.