files-view: Show and hide menu items programatically
As we've added more views and modes, we've sliced and diced the menu's accordingly.
It's becoming complicated and difficult to understand.  Instead, introduce some
filter functions to allow declaration in the ui code of whether a menu item
should be shown in a particular view or mode.

This commit also fixes a bug where the scripts menu is shown in the network
view