window-slot: Drop .new_content_view pointer
We've been using two instance struct member pointers to keep both the
current view and the new one which is about to replace it. This was
necessary to prepare the new view before getting rid of the old one.

But now we won't do that anymore. Once a content view is created,
that's the only view the slot is every going to have.

Use .content_view in all cases.