previewer: Remember invoker slot
When users press the arrow keys while the previewer window has focus,
we emulate the same event on the files view.

However, we do this by taking the most recently focused window in the
NautilusApplication, which poses two problems:

* there is no warranty that the most recently focused NautilusWindow
  is the one which the previewer was launched from.
* FileChooser windows do not belong to NautilusApplication, so we are
  100% sure to get the wrong window in that case.

Instead of passing a GtkRoot window to show previewer, pass a slot
and keep a pointer to it around to reuse when emulating arrows back.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/work_items/3428