file-chooser: Allow drops to the bottom bar
There's issues with DnD into the view of nautilus because it
has a different action for the main window vs the file chooser.
Furthermore, these actions can sometimes be destructive so we
want to be careful.  Let's do something really simple and just
allow DnD to the new bottom bar, which does not exist in the
main nautilus.

I'm also working on a separate branch to show open windows
which should make DnD much less important in general.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3538