files-view: Show background menu in filechooser
But hide all of the items that shouldn't be shown.  Visible Columns is already shown
in the view menu.  New Folder is already shown in the toolbar, but out of muscle
memory from the regular view, and for ergonomics, show it in the background menu.
Select All has no other non-keyboard option