views: Don't sort directories first in Recent
Directories aren't meant to be shown in the Recent folder,
but it's possible they can appear.  gvfs only checks the
mimetype stored in the GBookmarkFile for the recent entry
and does not check the actual mimetype on the file.  This
generally shouldn't be a problem, unless the mimetype was
set incorrectly.  If folders slip in, there's currently a
bug where you can't even delete the recent entry [1]
and folders are appearing because of another bug due
to gtk no longer being the file chooser [2]

With stale folders possibly included, they'll take over
the whole visible portion of the Recent folder when the
directories-first setting is set.  Since this bug can
possible appear in other areas, let's minimize the impact.
Also, if we ever do decide to show folders in Recent,
it still makes most sense to have the most recent item
first.

[1] https://gitlab.gnome.org/GNOME/gvfs/-/merge_requests/228
[2] https://gitlab.gnome.org/GNOME/xdg-desktop-portal-gnome/-/merge_requests/177