file-chooser: Set name on construction
If you start a file chooser with an empty current_name, filename validator
will spew out warnings for a NULL filename because we aren't setting
the name on construction.  Fortunately this isn't a problem when the
name is set (and conflicts) because our on_change listener ends up
setting the name name. But in order to get rid of the warnings and to
be more reliable set the name on construction.