toolbar: Allow hiding view and history controls separately
In narrow mode for the normal nautilus window, we move both the history and
view controls to the bottom bar.  In the file chooser we aren't moving either
to the bottom bar.  In folders that have a search info popover (i.e. where
results aren't indexed) the nautilus-toolbar gets just long enough to
exceed the 360px width, which cuts off part of the file chooser.

It doesn't really make sense to move the history controls to the bottom because
it's busy already, but it seems natural to move the view controls to the same
place they are in the narrow normal nautilus view.  Split out the property
for history and view controls which allows us to move the view controls
in file chooser mode.

Related: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3554