window-slot: Simplify remembering view mode
We store the view mode changing to network view, and restore it when
leaving it.

Our goal with this is to make the view mode persist per slot, even if
the setting has been changed from another slot/window.

So, let's change the logic to reflect this goal. It's functionally
equivalent but simpler to reason about. Also add new safety checks to
ensure view_mode is only ever set to LIST or GRID.