list-base: Fix scroll to zoom
Initially we implemented scroll to zoom with a scroll controller using
the DISCRETE flat in e358c2343c4 ("general: Switch to GTK4"), but
that caused an issue where all scrolls were discrete (even without
the ctrl modifier held).  In order to address that issue, commit
86f0c7e75a7 ("list-base: Don't scroll when the ctrl key is held")
attempted to fix the issue by only setting the discrete flag on
::scroll-begin.  That works for trackpads that support that signal,
but many modern mice actually emit several scroll events for each
page of scroll (without emitting a scroll-begin/end, and thus almost
never would match the dy>1 or dy<-1.

Instead use a more robust solution where we ignore the individual
scroll events themselves but look for a page of scrolling in a short
amount of time (100ms seemed reasonable to me).

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3425