properties-window: Use fallback for deep counts
The new way to calculate deep counts doesn't work on gvfs backends.  Take
the existing logic, we had in directory-async, but in a much more simple
manner.  In particular remove the inode check that was done in order not
to duplicate the count for hard links.  That's not applicable to these
other backends, and the implementation for checking for duplicate
inodes, was slowing down the count significantly.