file-chooser: Reset new folder button properly
The new folder button hides after making the window smaller and
then larger again because the breakpoint is resetting to the
"original" state.  The original state is recording when the
breakpoint setter is created in the ui file, before the new folder
button property is set.  Create the setter after the property is
correctly initialized.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3571