ui: Add missing accel labels
Commit 9dbe8569c3 ("window-slot: Don't add accels to application")
and associated commits, removed the shortcuts from the window
and placed them onto the widgets themselves (i.e. the slot and view).
When we used gtk_application_set_accelerator() (indirectly), GTK
was actually adding a shortcut controller to the window itself.
That allowed the GtkMenuPopover to find the accelerator via
a lookup process in GtkActionMuxer because the window was an
ancestor of the menu.  Now, however, the shortcut is not an
ancestor of the menu and thus GTK doesn't find the accel. We probably
could patch GTK, but that would be much more complicated.  Instead
set the "accel" attribute on a couple menu items where this is applicable.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3562


(cherry picked from commit 8eaecf9a9e5bd4eb2978180630b3fab521281b61)