readme: Clarify xdg-user-dirs-gtk dependency
Nautilus has quasi-depended on xdg-user-dirs-gtk, but it
has largely gone unnoticed.  That is, until commit 5c2f55663ad
("placessidebar: Reorganize places"), which makes this dependency
very obvious.  Along with updating localization, xdg-user-dirs-gtk
also creates the default bookmarks.  Now that we are treating
the XDG locations as regular bookmarks it's critical that
the user has xdg-user-dirs-gtk installed to create the default
bookmarks.  Ideally, the dependency would be installed as part of
the base install to ensure that it is executed before the
bookmark file is created elsewhere, but that's out of our control.