nautilus-properties-window: Set use-markup=False in several rows
If the file/folder path has an ampersand, the path won't
be displayed. Set use-markup=False in the "Parent Folder",
"Link Target" and "Original Folder" rows to fix that.

Also, use this property in several other rows that show
information that isn't controlled by Nautilus to prevent
the problem mentioning above from happening. The only
exceptions are the "Others" row and the "Permissions"
rows.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3653