search-engine-simple: Use MountMonitor to cache FS info
Simple search engine will check if directories are for a remote
filesystem before recursion for performance reasons. But when most
directories are local, checking will take the majority of the simple
search engine's CPU time since it's blocking the search thread on IO.

Internally, GIO will call cache time given by
g_unix_mount_entry_at(), but that requires the presence of a
GUnixMountMonitor to actually work [1]. So retain a reference to an
instance of GUnixMountMonitor in order for the cache to start
working again and eleminate almost all of the overhead of querying
filesystem info.

[1] https://gitlab.gnome.org/GNOME/glib/-/commit/05afb30510fb98d8a564759556d11f7a1037df5a

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3702


(cherry picked from commit e740773b41748300de8fa4df89b7136151487d75)

Co-authored-by: Khalid Abu Shawarib <khalid.shawarib@gmail.com>