Set the same height and width for all icons regardless of aspect ratio
Make the code do what the comment says.

The way the code was set up previously, it was setting the size on only
the "fixed_height_box", which contained both the icon and an additional
column for the emblems. As a result, if the icon was portrait aspect,
the box would end up shrinking horizontally until the combined icon +
emblems equaled the height of the icon.

The simplest solution is to set the size on the AdwClamp which directly
contains the GtkPicture. This ensures the size of the "fixed_height_box"
doesn't depend on the aspect ratio of the icon.

Note that the size still has to be set on the fixed_height_box to ensure
that labels stay properly aligned, particularly if some wrap to multiple
lines.