general: Split out thumbnail info from base file info fetching
We currently have "thumbnail::*" file attributes in the base set of loaded
info, and thus loading them unconditionally, even when showing thumbnails
is disabled in the settings (we load it but not show it in that case).

Moreover, we don't even begin to show the content unless all thumbnail
info of all files are loaded, since they are part of the base info. This
is the biggest factor in increasing loading time as observed in [1] and
[2].

This patch splits out thumbnail info loading from the base file info,
into a separate THUMBNAIL_INFO request, effectively allowing for the
load of the view without waiting for thumbail info. This can brings
substantial loading time reduction for large folders, especially ones
with lots of thumbnails.

[1] https://gitlab.gnome.org/GNOME/nautilus/-/issues/3374
[2] https://gitlab.gnome.org/GNOME/gtk/-/issues/2177