batch-rename-dialog: Make UI rudimentarily adaptive
This changes the list's layout from one row to two rows.

Besides the list, this sticks as as close as possible to
the old UI but tweaks the margins and spacings a bit.
Most notably, this replaces hard width requests of
entries with Adw.Clamps, and uses a breakpoint to switch
boxes' orientation for smaller widths.

While at it, it replaces Gtk.Grids with Gtk.Boxes where
the layout was only in one direction as a clean-up.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>