properties-window: Create dialog directly
Most commonly all files for a properties dialog are already loaded
and ready. This thus changes the dialog to always be created directly
instead of catering to the rare use-case of needing to wait for files
by delaying the dialog creation. Should some files not be ready yet,
an empty page with a spinner is shown.

Code-wise this removes the StartupData struct along with its many
helper functions and adds a new direct constructor instead. The
spinner-only initial state requires "content-height" to be set.