Revert "date-utilities: Remove uselocale workaround"
This reverts commit 727175704bfa6dd3188b235cff35a482d9cc1ab7.

Turns out this will only work if we ship translations in the LC_TIME
folder "localedir/LC_TIME/" instead of "localedir/LC_MESSAGES".

No app on my system does this, and gettext docs asks you not to do it [1].
Moreover, it seems like the solution employed by glib [2] is basically
the same as the one previously used.

[1] https://www.gnu.org/software/libc/manual//2.28/html_mono/libc.html#Using-gettextized-software
[2] https://gitlab.gnome.org/GNOME/glib/-/blob/05da47c2d0d333cec6363e09bcf9eed6952e74ce/gio/gsettingsschema.c?page=2#L1442-1451

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3791