grid-cell: Manage our own layout
Instead of abusing AdwClamp to prevent the icon and the labels from
expanding beyond their minimum size, and using a "fixed_height_box" to
force a part of the layout to have a fixed height, take the layout into
our own hands.

We can both return exactly the desired size from our
nautilus_grid_cell_measure () implementation, while also making the cell
constant-size, and place and allocate sizes to the icon and the labels
exactly how we want to.

This both lets us control the layout a lot more directly, makes things a
lot simpler and more efficient for the GTK layout machinery, and is very
unlikely to break.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>