files-view: Don't wait for attributes to load directory
This restores original behaviour before [1] which was there for some
reason related to how it conflicts with the monitoring of the directory.

I can't find any issue with it restored back neither in main or in 48,
so might've been a fluke or some temporary issue with code as written at
the time of testing.

[1] https://gitlab.gnome.org/GNOME/nautilus/-/commit/fd8fb982a64b24022cba3744d084dabb4b76f26f

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3834