application: Register to xdg-desktop-portal
Currently, trying to use xdg-desktop-portal results in problems with
missing app id, i.e. when trying to inhibit logout during file copying.

GTK started prefering xdg-desktop-portal even in host apps, which
obtains app ids from the systemd unit name. However it only works for
the ones adhering to xdg standard scheme, which currently is not the
case for dbus activated apps. During GtkApplication startup there is a
call to the register API, but it needs to be the first call to the
portal for it to work. Unfortunately, we indirectly interact with the
portal before that, when initializing GDK display before the chain-up,
which is necessary for proper functioning of the file chooser mode.

Fix that by registering with the xdg-desktop-portal before initizalizing
the GDK display.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/3874