extensions: Make property docs proper GIR comments
The properties of `NautilusMenuItem` were documented in a pretty random
C comment. This commit moves to a more appropriate place in the .c file
and migrates them to a GIR-formatted content, which makes them show up
in the reference documentation.

As part of this change, we're also doing a little cleanup where we use a
more modern construction with `g_object_class_install_properties()`.