Remove superflous assertion which leads to crash
Assertion filename[0] != '' in nautilus_file_new_from_filename()
and nautilus_directory_new_file_from_filename() is
superflous because while it asserts on empty filename
it's otherwise perfectly fine with eg. a one space ' '
filename i.e. it does not crash in this case.

I traced back the introduction of the filename[0]
assertion just in case it was due to some specific
bug, but nothing to do with that, it originated[1]
from 24 years ago with initial implementations of
functions.

Fixes #3862

[1] https://gitlab.gnome.org/GNOME/nautilus/-/commit/e1f2a77d0879e58c20b1bbc5ef183ccbce16bdb6#68dce43a5e34629fdf3a9febcfb931d30a4a3026_118_119