search-engine: Stop acting like a search provider
This is in preparation to turn search provider from an interface into a
derivable class. It also clarifies that there is a conceptual
difference between the main search engine and search provider engines.

Use the opportunity to turn signals into callbacks, as signals can have
multiple subscribers and transferring of search hits is thus not great
to do via a signal.