search-engine: Stop acting like a search provider
This clarifies that there is a conceptual difference between the main search
engine and search provider engines. It also is in preparation for adding run
IDs to the provider start function.

Use the opportunity to combine "finished" and "error" into a
"search-finished" signal and provide an optional error message with it.