search-popover: Present date ranges as tag cloud
This modernizes the date range selection in the search popover by
providing a tag cloud to replace the previous nested list. It matches
the latest mockup at [1] closely, but uses different default date
ranges.

It removes the capability to choose which file date type to search for,
as well as the capability to choose dates from a calendar. These will be
brought back in future commits.

[1] https://gitlab.gnome.org/Teams/Design/app-mockups/-/blob/master/files/files-search-filters-aday.png

Co-authored-by: Anuraag Reddy Patllollu <anuraagreddy74@gmail.com>