Add DateRangeDialog
This dialog allows to select a date range by using two of the
previously introduced DateEntryRows. If the Select button is
pressed a date range is created which can be queried by
subscribing to the "close-attempt" signal. If the dialog
is simply closed, or somehow has an invalid date range, the
date range is NULL.