application: Use startup_id when presenting windows
The startup ID that was passed in via the FileManager1 DBus interface
was dropped when nautilus_application_open_location() was calling
nautilus_application_open_location_full(). This meant that if a window
was being presented, it was not using the correct token/startup id,
which would cause the focus stealing prevention in some window managers,
such as mutter, to refuse focusing the window.

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3385