search-engine: Simplify restart logic
As search engine threads run in the same context, we can rely on the
restart variable alone. Make the logic more obvious by setting it
together with self-referencing in the appropriate places. Also, only
consider the engine really stopped when all providers are done, not
directly after calling the stop function.