directory: Fix a crash when moving files
Regression caused by a combination of [1] and [2], where a file is
freed by [1] even though it is put in the changed_lists hash table,
which collect files to emit "changed" signals. The issue later got
exposed by [2].

Add it to the unref_list instead which is freed after emission of
signals.

[1] 483aabfcbfd4552982b93f3fe2420908a9646d93
[2] 59af3eac46f34806106470c308c2640a5275ee90

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3920