files-view: Don't schedule context menu update before processing changes
This call to update context menu is tied directly to the directory's
files-changed signal, the content of which has not yet been filtered
or processed. Therefore, it should be called when processing the
pending files instead, and only consider changes concerning selection
as those are the ones that affect menus.