test/files-view: Assert renames are subset of changes instead of equality
With [1] and [2], the "files-changed" signal emittion is polluted with
changes to the thumbnail info status that are making their way into
files-view tests listeners for that signal, and thus can't rely on specific
changes such as renaming to be exactly the same as the files collected from
"files-changed" signal listeners.

Best we can do now is to check if it is a subset of changes rather than
checking if it is exactly equal to the list files that we expect to have
changed.

[1] 97b7f734d2aa8841f29bed9955dd08dca06b8c0b
[2] 31e8fdc8492391958befe52287a1f112b5c84ddb