view: Show files in MTP-backed directories incrementally while loading
If there are lots of files in a directory accessed via MTPm we currently wait
for all the files in the folder to be enumerated before showing any files.
As the comment here already mentions, this makes sense for local directories,
but is an issue with remote directories, as here the wait for the directory to
fully load can be several minutes.

There already is code in place for incremental loading of search results, so
just adapt the if-statement here and prematurely show the files also in case
the folder location is an MTP device.

Co-Authored-By: re2zero <yangwu@uniontech.com>

view: Show files in MTP-backed directories incrementally while loading

If there are lots of files in a directory accessed via MTPm we currently wait
for all the files in the folder to be enumerated before showing any files.
As the comment here already mentions, this makes sense for local directories,
but is an issue with remote directories, as here the wait for the directory to
fully load can be several minutes.

There already is code in place for incremental loading of search results, so
just adapt the if-statement here and prematurely show the files also in case
the folder location is an MTP device.

Co-Authored-By: re2zero <yangwu@uniontech.com>