name-cell: Fix uneven margin
When the paintable height is an odd number, the extra_margin var gets
truncated from a fraction before being used for as a value for setting
the margin. This causes a loss of 1 pixel in overall size.

We store the result in a float instead to retain the fraction when
halfing. But since widget margin setter only accepts integers, the extra
pixel has to be given to one of the margin unevenly. This is done through
alernating between ceil() and floor() for the margin values.

This solves an issue with 1 pixel size delta between cut items and normal
ones. It is due to cut icons being fixed at even sizes, but images might
have odd paintable size.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/4007