properties-window: Don't depend on icon being already available
It was asking for file icons with thumbnail flags in the function
get_image_for_properties_window(), but it didn't request them in the
async file attribute query. That made it depend on the view to get
the icons.

Perform a proper query for thumbnails before asking for file
thumbnails.