general: Always try packagekit option
Instead of hiding this 17+ years old feature to offer installation of
apps via packagekit behind an option, simply always try to make the DBus
call and fallback if it fails.

Also drop the user setting for "admins" to disable this, as with Flatpak
apps can be installed in user space.