directory-async: Move callbacks to separate struct
This separates callbacks from NautilusDirectoryPrivate by using a
dedicated struct for them. This allows for more clearer memory
management, as it is now handled in dedicated functions, that are called
respectively in init/finalize of NautilusDirectory.

This is in preparation for changing the underlying callback storage data
structure in upcoming commits.