directory: Split notifying and work queue adding
In [1] the directory check in the call_files_changed_common helper
function was dropped under the assumpiton that the passed directory will
always match the parameter. This is not the case when moving, where the
passed directory can be the directory an item was moved from. In this
case the file should not be added to the directories work queue, as it
is no longer contained in it and can't be processed.

As a solution split the function up into a work queue part and a
notifying part. This avoids going back to the previous extra check that
itself was added as a fix in [2] over 20 years ago.

[1] https://gitlab.gnome.org/GNOME/nautilus/-/commit/f9231491d4a851f68f18e40048679bf79ec43d95
[2] https://gitlab.gnome.org/GNOME/nautilus/-/commit/07cc9c27cd0cfaf0741d18103d779b4474dced0a